made code more modular included init.py simplified main.py
