+++ project_bootstrapper/__init__.py
@@ -1,0 +1,17 @@
+from .config import prompt_for_valid_directory, load_config
+from .project_data_templates import DATA_SCIENCE_FOLDERS, DATA_SCIENCE_FILES, get_lightweight_files, get_lightweight_folders, GITIGNORE_ENTRIES
+from .scaffold import create_structure
+from uv_init import dependency_check, init_project
+
+__all__ = [
+    "prompt_for_valid_directory",
+    "load_config",
+    "DATA_SCIENCE_FOLDERS",
+    "DATA_SCIENCE_FILES",
+    "get_lightweight_files",
+    "get_lightweight_folders",
+    "GITIGNORE_ENTRIES",
+    "create_structure",
+    "dependency_check",
+    "init_project"
+]
